+++ core/modules/settings/class-network.php
@@ -11,6 +11,21 @@ defined( 'ABSPATH' ) || exit;
  * @author Varun Sridharan <varunsridharan23@gmail.com>
  */
 class Network extends Settings {
+
+	/**
+	 * Stores Module DB Type.
+	 *
+	 * @var string
+	 */
+	protected $module_db = 'network_settings';
+
+	/**
+	 * Module Type.
+	 *
+	 * @var string
+	 */
+	protected $module = 'network_settings';
+
 	/**
 	 * Inits Class Instance.
 	 *
@@ -42,7 +57,7 @@ class Network extends Settings {
 	 */
 	public function on_settings_page_load() {
 		if ( isset( $_POST[ $this->unique ] ) ) {
-			$this->set_db_values( $this->save_validate( $_POST[ $this->unique ] ) );
+			$this->save_validate( $_POST[ $this->unique ] );
 		}
 		parent::on_settings_page_load();
 	}