+++ CHANGELOG.md
@@ -5,6 +5,37 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [1.5.3.2] - 23/07/2020
+## Added
+* DB Save Hook Support For Settings & Network Settings Module
+* `set_unique` function in **Unique** Trait
+* `set_module` function in **Module** Trait
+* new `object` Trait created
+* Custom Save Handler For Every Module.
+* **Action :** `wponion/metabox/render/before`
+* **Action :** `wponion/metabox/render/{$this->unique()}/before`
+* **Action :** `wponion/metabox/render/{$this->unique()}/after`
+* **Action :** `wponion/metabox/render/after`
+* Custom Class Option In Admin Page Module
+* Notification Bubble Support In Admin Page Module
+* Menu Separator Support In Admin Page Module
+ 
+## Fixed
+* Taxonmy Register Issue . fixed by changing `$this->taxonomy` to `$this->unique()`
+* **CSS** Tab layout `z-index`
+* Ajax issue with metabox saving
+
+## Changed
+* Bump [@wordpress/hooks] from `2.8.0` to `2.9.0`
+* Bump [@popperjs/core] from `2.4.2` to `2.4.4`
+* Bump [tippy.js] from `6.2.3` to `6.2.5`
+* Bump [sweetalert2] from `9.15.1` to `9.17.0`
+* Bump [css-checkbox-library] from `1.5.1` to `1.5.2`
+
+## Removed
+* Deleted local function calls added by mistake
+
+
 # [1.5.3.1] - 23/06/2020
 ## Fixed
 * Custom Template Tags Not Working In Selectbox
@@ -952,6 +983,7 @@ Migrated From WordPress VIP Coding Standards To WordPress Core Coding Standards.
 8. WP Module   -  Settings Page
 
 ---
+[1.5.3.2]: https://github.com/wponion/wponion/releases/tag/1.5.3.2
 [1.5.3.1]: https://github.com/wponion/wponion/releases/tag/1.5.3.1
 [1.5.3]: https://github.com/wponion/wponion/releases/tag/1.5.3
 [1.5.2]: https://github.com/wponion/wponion/releases/tag/1.5.2